requirements.txt
Django==5.2.4
djangorestframework==3.14.0
drf-yasg==1.22.1
django-filter==23.2
gunicorn==20.1.0
psycopg2-binary==2.9.6
whitenoise==6.5.0
PyPDF2==3.0.1
django-cors-headers==3.14.0
django-environ==0.10.0